Work in progress...
I really do not know what i did here. It seems we have two working
builds in place which is nice. And there is a unfinished udp server
class here. Its called async server. I forgot why i named it async.
Did i want to use nio? Or Should it just be called UdpServer. I commit
and push this because i want to work on in on my laptop. This repo seems
to evolve to a prototype.